Move send_domain_registration_email into registration/tasks
