+++ corehq/apps/registration/tasks.py
@@ -1,5 +1,6 @@
 from __future__ import absolute_import
 from __future__ import unicode_literals
+import logging
 from celery.schedules import crontab
 from celery.task import periodic_task
 from django.conf import settings
@@ -50,3 +51,34 @@ def activation_24hr_reminder_email():
             text_content=message_plaintext,
             email_from=settings.DEFAULT_FROM_EMAIL
         )
+
+
+WIKI_LINK = 'http://help.commcarehq.org'
+FORUM_LINK = 'https://forum.dimagi.com/'
+PRICING_LINK = 'https://www.commcarehq.org/pricing'
+
+
+def send_domain_registration_email(recipient, domain_name, guid, full_name, first_name):
+    DNS_name = get_site_domain()
+    registration_link = 'http://' + DNS_name + reverse('registration_confirm_domain') + guid + '/'
+    params = {
+        "domain": domain_name,
+        "pricing_link": PRICING_LINK,
+        "registration_link": registration_link,
+        "full_name": full_name,
+        "first_name": first_name,
+        "forum_link": FORUM_LINK,
+        "wiki_link": WIKI_LINK,
+        'url_prefix': '' if settings.STATIC_CDN else 'http://' + DNS_name,
+    }
+    message_plaintext = render_to_string('registration/email/confirm_account.txt', params)
+    message_html = render_to_string('registration/email/confirm_account.html', params)
+
+    subject = ugettext('Activate your CommCare project')
+
+    try:
+        send_html_email_async.delay(subject, recipient, message_html,
+                                    text_content=message_plaintext,
+                                    email_from=settings.DEFAULT_FROM_EMAIL)
+    except Exception:
+        logging.warning("Can't send email, but the message was:\n%s" % message_plaintext)
