+++ corehq/apps/registration/utils.py
@@ -6,7 +6,6 @@ from django.utils.translation import ugettext
 import uuid
 from datetime import datetime, date, timedelta
 from django.template.loader import render_to_string
-from celery.task import task
 from corehq.apps.accounting.models import (
     SoftwarePlanEdition, DefaultProductPlan, BillingAccount, BillingContactInfo,
     BillingAccountType, Subscription, SubscriptionAdjustmentMethod, Currency,
@@ -16,6 +15,7 @@ from corehq.apps.accounting.models import (
 from corehq.apps.accounting.tasks import ensure_explicit_community_subscription
 from corehq.apps.app_manager.tasks import load_appcues_template_apps
 from corehq.apps.registration.models import RegistrationRequest
+from corehq.apps.registration.tasks import send_domain_registration_email
 from dimagi.utils.couch import CriticalSection
 from dimagi.utils.name_to_url import name_to_url
 from dimagi.utils.web import get_ip, get_url_base, get_site_domain
@@ -154,37 +154,6 @@ def request_new_domain(request, form, is_new_user=True):
 
     send_hubspot_form(HUBSPOT_CREATED_NEW_PROJECT_SPACE_FORM_ID, request)
     return new_domain.name
-
-
-WIKI_LINK = 'http://help.commcarehq.org'
-FORUM_LINK = 'https://forum.dimagi.com/'
-PRICING_LINK = 'https://www.commcarehq.org/pricing'
-
-
-def send_domain_registration_email(recipient, domain_name, guid, full_name, first_name):
-    DNS_name = get_site_domain()
-    registration_link = 'http://' + DNS_name + reverse('registration_confirm_domain') + guid + '/'
-    params = {
-        "domain": domain_name,
-        "pricing_link": PRICING_LINK,
-        "registration_link": registration_link,
-        "full_name": full_name,
-        "first_name": first_name,
-        "forum_link": FORUM_LINK,
-        "wiki_link": WIKI_LINK,
-        'url_prefix': '' if settings.STATIC_CDN else 'http://' + DNS_name,
-    }
-    message_plaintext = render_to_string('registration/email/confirm_account.txt', params)
-    message_html = render_to_string('registration/email/confirm_account.html', params)
-
-    subject = ugettext('Activate your CommCare project')
-
-    try:
-        send_html_email_async.delay(subject, recipient, message_html,
-                                    text_content=message_plaintext,
-                                    email_from=settings.DEFAULT_FROM_EMAIL)
-    except Exception:
-        logging.warning("Can't send email, but the message was:\n%s" % message_plaintext)
 
 
 def send_new_request_update_email(user, requesting_ip, entity_name, entity_type="domain", is_new_user=False, is_confirming=False):
